// src/api/types.ts
export type TaskStatus = 'pending' | 'in-progress' | 'completed'
export type TaskPriority = 'low' | 'medium' | 'high'

export interface Task {
  id: string
  title: string
  description: string
  status: TaskStatus
  priority: TaskPriority
  createdAt: string
}

export type CreateTaskDto = Pick<Task, 'title' | 'description' | 'status' | 'priority'>

export type UpdateTaskDto = Partial<Pick<Task, 'title' | 'description' | 'status' | 'priority'>>